Update distance.py and move module from main.py to distance.py
